Allows quick selection of last result used while importing library
