+++ fe/src/components/domain/audiobook/LibraryImportSearchModal.vue
@@ -31,27 +31,12 @@
         </div>
 
         <div v-if="searchResults.length > 0" class="results-list">
-          <div
+          <SearchResultComponent
             v-for="result in searchResults"
-            :key="result.asin ?? result.title"
-            class="result-item"
+            :result="result"
+            :placeholder-url="placeholderUrl"
             @click="select(result)"
-          >
-            <img
-              v-if="result.imageUrl"
-              :src="getProtectedImageSrc(result.imageUrl, `library-import-search-${result.asin ?? result.title}`, placeholderUrl)"
-              class="result-thumb"
-              alt=""
-            />
-            <div class="result-info">
-              <span class="result-title">{{ result.title }}</span>
-              <span class="result-meta">
-                {{ result.authors?.[0]?.name }}
-                <span v-if="result.series"> · {{ Array.isArray(result.series) ? (result.series as any)[0]?.name : result.series }}</span>
-                <span v-if="result.asin" class="result-asin"> · {{ result.asin }}</span>
-              </span>
-            </div>
-          </div>
+          />
         </div>
 
         <div v-else-if="hasSearched && !isSearching" class="no-results">
@@ -60,6 +45,17 @@
 
         <div v-else-if="!hasSearched && !isSearching" class="hint-text">
           Type a title or paste an ASIN to search
+        </div>
+
+        <div v-if="lastSelected != null">
+          <span>Quick select:</span>
+          <div class="results-list">
+            <SearchResultComponent
+              :result="lastSelected"
+              :placeholder-url="placeholderUrl"
+              @click="select(lastSelected)"
+            />
+          </div>
         </div>
       </div>
     </ModalBody>
@@ -71,19 +67,21 @@ import { ref, onMounted, nextTick } from 'vue'
 import { PhSpinner } from '@phosphor-icons/vue'
 import { Modal, ModalHeader, ModalBody } from '@/components/feedback'
 import { apiService } from '@/services/api'
-import { useProtectedImages } from '@/composables/useProtectedImages'
 import type { LibraryImportItem } from '@/stores/libraryImport'
 import { buildLibraryImportInitialAuthor, buildLibraryImportInitialQuery } from '@/utils/libraryImportSearch'
 import { getPlaceholderUrl } from '@/utils/placeholder'
 import type { SearchResult } from '@/types'
+import SearchResultComponent from './SearchResultComponent.vue'
 
-const props = defineProps<{ item: LibraryImportItem }>()
+const props = defineProps<{ 
+  item: LibraryImportItem, 
+  lastSelected: SearchResult | null 
+}>()
 const emit = defineEmits<{
   close: []
   select: [result: SearchResult]
 }>()
 
-const { getProtectedImageSrc } = useProtectedImages()
 const inputEl = ref<HTMLInputElement | null>(null)
 const placeholderUrl = getPlaceholderUrl()
 // Build the initial query: ASIN → filename stem (when more specific than folder) → folderName
@@ -167,60 +165,6 @@ function select(result: SearchResult) {
   overflow-y: auto;
   border: 1px solid #333;
   border-radius: 6px;
-}
-
-.result-item {
-  display: flex;
-  align-items: center;
-  gap: 0.6rem;
-  padding: 0.5rem 0.75rem;
-  cursor: pointer;
-  transition: background 0.15s;
-  border-bottom: 1px solid #2a2a2a;
-}
-
-.result-item:last-child {
-  border-bottom: none;
-}
-
-.result-item:hover {
-  background: #2a2a2a;
-}
-
-.result-thumb {
-  width: 36px;
-  height: 36px;
-  object-fit: cover;
-  border-radius: 3px;
-  flex-shrink: 0;
-}
-
-.result-info {
-  min-width: 0;
-  flex: 1;
-}
-
-.result-title {
-  display: block;
-  font-size: 0.875rem;
-  color: #e0e0e0;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.result-meta {
-  display: block;
-  font-size: 0.75rem;
-  color: #888;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.result-asin {
-  font-family: monospace;
-  font-size: 0.7rem;
 }
 
 .no-results,
